test: relocate mocks
