+++ test/src/commands/rm/rm_test.dart
@@ -5,7 +5,7 @@ import 'package:coverde/src/commands/rm/rm.dart';
 import 'package:mocktail/mocktail.dart';
 import 'package:test/test.dart';
 
-import '../../utils/mocks.dart';
+import '../../../utils/mocks.dart';
 
 void main() {
   group(
